Reduce plmn postprocessor false positives with dual SIM and MOCN
When there are two SIMs connected to two cells with the same channel but different MNC, the postprocessor copies the PLMN of the first cell to the second, but this could be wrong if the two SIMs access the same network with different MNC (MOCN).
This change avoids this problem by not changing the PLMN if there're multiple primary cells with the same channel.
Also added a test for this case